+++ include/matrix/matrix_float.h
@@ -5,6 +5,6 @@ void print_float(void* element);
 void add_float(void* result, void* a, void* b);
 void mult_float(void* result, void* a, void* b);
 void mult_float_const(void* result, void* a, int b);
-Matrix* create_float_matrix(size_t razm);
+Matrix* create_float_matrix(int razm);
 
 #endif //MATRIX_FLOAT_H
