+++ src/matrix.c
@@ -3,7 +3,7 @@
 #include "../include/matrix/matrix.h"
 #include "../include/func.h"
 
-Matrix* matrix_create(size_t razm, size_t element_size,
+Matrix* matrix_create(int razm, size_t element_size,
                       void (*print_element)(void*),
                       void (*add_elements)(void*, void*, void*),
                       void (*multiply_elements)(void*, void*, void*),
